Small refactor to move the link generation out of the view, to make it more reusable from Python
