+++ wagtaildraftsharing/tests/test_models.py
@@ -1,11 +1,71 @@
+import datetime
 from textwrap import dedent
 from unittest.mock import patch
 
 import wagtail
+from django.contrib.auth.models import User
 from django.test import TestCase
+from django.utils.timezone import is_aware
+from freezegun import freeze_time
 from wagtail_factories import PageFactory
 
-from ..models import WagtaildraftsharingLink
+import wagtaildraftsharing.models
+from wagtaildraftsharing.models import WagtaildraftsharingLink
+
+FROZEN_TIME_ISOFORMATTED = "2024-01-02 12:34:56.123456+00:00"
+
+
+class TestWagtaildraftsharingLinkManager(TestCase):
+    def setUp(self):
+        self.test_user = User.objects.create(
+            username="test", email="testuser@example.com"
+        )
+
+    def create_revision(self):
+        page = PageFactory()
+
+        # create the first revision
+        page.save_revision().publish()
+
+        old_title = page.title
+        new_title = f"New {old_title}"
+        page.title = new_title
+
+        # create the second revision with a new title
+        page.save_revision().publish()
+
+        page.refresh_from_db()
+        earliest_revision = page.revisions.earliest("created_at")
+        return earliest_revision
+
+    @freeze_time(FROZEN_TIME_ISOFORMATTED)
+    def test_create_sharing_link_view__max_age_from_settings(self):
+        frozen_time = datetime.datetime.fromisoformat(FROZEN_TIME_ISOFORMATTED)
+
+        # Ensure we've got a level playing field: that the time is TZ-aware
+        if not is_aware(frozen_time):
+            self.fail("frozen_time was a naive datetime but it should not be")
+
+        max_ages_and_expected_expiries = (
+            (300, frozen_time + datetime.timedelta(seconds=300)),
+            (1250000, frozen_time + datetime.timedelta(seconds=1250000)),
+            (-1, None),
+        )
+
+        for max_age, expected_expiry in max_ages_and_expected_expiries:
+            with self.subTest(max_age=max_age, expected_expiry=expected_expiry):
+                with patch.object(wagtaildraftsharing.models, "max_age", max_age):
+                    revision = self.create_revision()
+
+                    link = WagtaildraftsharingLink.objects.get_or_create_for_revision(
+                        revision=revision,
+                        user=self.test_user,
+                    )
+
+                    assert link.active_until == expected_expiry, (
+                        link.active_until,
+                        expected_expiry,
+                    )
 
 
 class TestWagtaildraftsharingLinkModel(TestCase):
